requirements.txt
# Core ML and data processing
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Agent and API communication
requests>=2.28.0
httpx>=0.24.0
websockets>=11.0.0

# Data visualization and analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter for exploration
jupyter>=1.0.0
ipykernel>=6.0.0

# Logging and configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Progress tracking
tqdm>=4.64.0

# Metrics and evaluation
tensorboard>=2.10.0
wandb>=0.13.0

# Utilities
click>=8.0.0
rich>=12.0.0

# LangGraph and LangChain for agentic workflows
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

# Authentication
globus-sdk>=3.0.0 